[vertexai] File for Schema integration tests
